fix addDomain modal && implement redirect to dashboard if auth or /login if not auth
